project_04_learning_journal/requirements.txt
awesome-slugify==1.6.5
bcrypt==3.1.3
cffi==1.11.1
click==6.7
Flask==0.12.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.0
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
peewee==2.10.1
pycparser==2.18
regex==2017.9.23
six==1.11.0
Unidecode==0.4.21
Werkzeug==0.12.2
WTForms==2.1
